Moved WebcastDropdown to separate file; added props validation
